requirements.txt
opencv-python
numpy==1.16.4
Pillow